@@ -1,21 +1,26 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.3.10'
-    ext.support_library_version = '28.0.0'
-    ext.dokka_version = '0.9.17'
-    ext.buildtools_version = '28.0.3'
+    ext{
+        gradle_plugin_version = '4.1.1'
+        kotlin_version = '1.4.10'
+        support_library_version = '28.0.0'
+        dokka_version = '0.9.17'
+        buildtools_version = '28.0.3'
+        jfrog_version='4.15.1'
+    }
+
     repositories {
         jcenter()
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath "com.android.tools.build:gradle:$gradle_plugin_version"
         classpath "org.jetbrains.dokka:dokka-android-gradle-plugin:$dokka_version"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-
-        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4'
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
+        classpath "org.jfrog.buildinfo:build-info-extractor-gradle:$jfrog_version"
+//        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4'
+//        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
     }
 }
 
@@ -28,4 +33,17 @@ allprojects {
 
 task clean(type: Delete) {
     delete rootProject.buildDir
+}
+
+ext {
+    build_tool_version = '30.0.2'
+    build_sdk_version = 30
+    target_api_level = 30
+    sdk_min_api_level = 17
+    min_api_level = 17
+    support_version = '28.+'
+
+    artifactory_user='justin'
+    artifactory_password='123456'
+    artifactory_contextUrl='http://192.168.50.6:8081/artifactory/'
 }
