@@ -1,15 +1,17 @@
 package com.arasthel.spannedgridlayoutmanager.sample
 
 import android.os.Bundle
-import android.support.v7.widget.RecyclerView
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.RecyclerView
 import com.arasthel.spannedgridlayoutmanager.SpanSize
 import com.arasthel.spannedgridlayoutmanager.SpannedGridLayoutManager
 import com.arasthel.spannedgridlayoutmanager.SpannedGridLayoutManager.Orientation.*
+import kotlin.random.Random
 
 /**
  * Created by Jorge Martín on 24/5/17.
  */
-class MainActivity: android.support.v7.app.AppCompatActivity() {
+class MainActivity: AppCompatActivity() {
 
     val recyclerview: RecyclerView by lazy { findViewById<RecyclerView>(R.id.recyclerView) }
 
@@ -18,7 +20,7 @@ class MainActivity: android.support.v7.app.AppCompatActivity() {
 
         setContentView(R.layout.activity_main)
 
-        val spannedGridLayoutManager = SpannedGridLayoutManager(orientation = VERTICAL, spans = 4)
+        val spannedGridLayoutManager = SpannedGridLayoutManager(orientation = HORIZONTAL, spans = 2)
         spannedGridLayoutManager.itemOrderIsStable = true
 
         recyclerview.layoutManager = spannedGridLayoutManager
@@ -34,11 +36,20 @@ class MainActivity: android.support.v7.app.AppCompatActivity() {
         }
 
         spannedGridLayoutManager.spanSizeLookup = SpannedGridLayoutManager.SpanSizeLookup { position ->
-            if (adapter.clickedItems[position]) {
-                SpanSize(2, 2)
-            } else {
-                SpanSize(1, 1)
-            }
+//            if (adapter.clickedItems[position]) {
+//                SpanSize(Random.nextInt(adapter.itemCount)%3 + 1, Random.nextInt(adapter.itemCount)%3 + 1)
+//            } else {
+//                SpanSize(1, 1)
+//            }
+             if(position % 3 == 0){
+                 SpanSize(1, 2)
+             }else  if(position % 3 == 1){
+                 SpanSize(2, 1)
+             }else  if(position % 3 == 2){
+                 SpanSize(2, 2)
+             }else{
+                 SpanSize(1, 1)
+             }
         }
 
         recyclerview.adapter = adapter
